+++ src/release_notes.md
@@ -1,4 +1,4 @@
-# Release v4.9.3.0
+# Release v4.9.4.0
 
 ---
 
@@ -6,7 +6,13 @@
 
 ### CLI
 
-- Fixed `css -update` problem when trying to update itself locked by the current process
+- #406: Custom DLL (CommandGathering.dll) Fails Compile-Time Namespace Resolution with //css_ref
+- Improved `css -cmd` UX when generating custom commands descriptions.
+- Updated custom commands `-pkill` to allow killing the process by PID.
+- Added custom commands:
+  - `-runas`
+  - `-which`
+  - `-who`
 
 ### CSScriptLib